2017/11/05 12:39 KafkaSingelClient KafkaSingelServer BenchMark来测试FileChannel SingleTransferTo性能

+++ src/main/java/com/basic/benchmark/task/ReceiverTask.java
@@ -23,7 +23,7 @@ public class ReceiverTask implements Runnable {
     @Override
     public void run() {
         ByteBuffer buf = ByteBuffer.allocate(1024*1024);
-        int length=0;
+        long length=0;
         try {
             while(socketChannel.read(buf) != -1){
                 buf.flip();
